backend principal built and methods fixed

+++ backend/src/main/java/com/example/demo/todo/TodoService.java
@@ -3,8 +3,8 @@ package com.example.demo.todo;
 import com.example.demo.todo.model.Todo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.Collection;
+;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -17,21 +17,25 @@ public class TodoService {
         return todoRepository.save(todo);
     }
 
-    public Collection<Todo> findAll() {
-        return todoRepository.findAll();
+    public List<Todo> getTodosbyUserEmail(String email) {
+        return todoRepository.findAllByUserEmail(email);
     }
 
     public Optional<Todo> findById(String id) {
             return todoRepository.findById(id);
         }
-    public Optional<Todo> updateTodo (String id, Todo updatedTodo) {
-        return todoRepository.findById(id)
+    public Optional<Todo> updateTodo (String id, String email, Todo updatedTodo) {
+        return todoRepository.findByIdAndUserEmail(id, email)
                 .map(todo -> todo.update(updatedTodo))
                 .map(todoRepository::save);
     }
 
-    public void deleteById(String id){
-        todoRepository.deleteById(id);
+    public Optional<Todo> deleteById (String id, String email){
+        Optional<Todo> todoToDelete = todoRepository.findByIdAndUserEmail(id, email);
+        if(todoToDelete.isPresent()) {
+            todoRepository.deleteById(id);
+        }
+        return todoToDelete;
     }
 
 }
